src/auth_server/requirements.txt
fastapi[standard]
sqlmodel
psycopg2-binary
pyjwt
bcrypt==4.0.1
passlib==1.7.4